Extract query formatting helpers
